requirements.txt
confluent-kafka==1.5.0
kafka-python==2.0.1
msgpack==1.0.0
PyYAML==5.3.1
